requirements.txt
tomli
requests
mutagen
pydub
pyloudnorm
pillow
ffmpeg-python
tqdm
numpy
soundfile
PlexAPI
musicbrainzngs
eyed3
tinytag
pyacoustid
fuzzywuzzy
Unidecode
python-magic
pyFLAC
m3u8
multiprocessing-logging
joblib
cryptography
keyring
matplotlib
seaborn
pandas
numba
rapidfuzz
syncthing
tidal-dl
youtube_dl
google-api-python-client
soundcloud-lib